backend/requirements.txt
# FastAPI core
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
sqlalchemy==2.0.40
asyncmy==0.2.10
aiosqlite==0.20.0
alembic==1.15.2

# Auth & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP client (for proxying APIs)
httpx==0.28.1

# Caching
cachetools==5.5.2

# Settings
pydantic-settings==2.8.1

# Payments
requests==2.32.3

# CORS
# (included in FastAPI)

# Dev / Testing
pytest==8.3.5
pytest-asyncio==0.25.3
